requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
paramiko==3.3.1
psutil==5.9.5
GitPython==3.1.37
APScheduler==3.10.4
redis==5.0.1
eventlet==0.33.3
gunicorn==21.2.0
